Move the code for creating new list into it's own component
