+++ src/views/Home.jsx
@@ -1,44 +1,10 @@
 import './Home.css';
 import { SingleList } from '../components';
-import { useState } from 'react';
-import { createList } from '../api';
-import { useNavigate } from 'react-router-dom';
-import toast, { Toaster } from 'react-hot-toast';
+import { CreateList } from '../components/CreateList';
 
 export function Home({ data, setListPath, userId, userEmail }) {
 	const hasList = data.length !== 0;
 
-	const [inputValue, setInputValue] = useState('');
-	const navigate = useNavigate();
-
-	const handleSubmit = async (e) => {
-		e.preventDefault();
-
-		// Ensure inputValue has a length > 0 after trimming
-		if (!inputValue.trim()) {
-			toast.error('List name cannot be empty.');
-			return;
-		}
-
-		try {
-			await createList(userId, userEmail, inputValue);
-			const path = `${user.uid}/${inputValue}`;
-			setListPath(path);
-			setInputValue('');
-			toast.success('Success: Your New List is Created!');
-			// Delay for toast notification before redirecting
-			setTimeout(() => {
-				navigate('/list');
-			}, 1500); // 1.5 seconds delay
-		} catch {
-			console.error('Error creating list:', error);
-			toast.error('Failed to create the list. Please try again.');
-		}
-	};
-
-	const handleChange = (e) => {
-		setInputValue(e.target.value);
-	};
 	return (
 		<>
 			<div className="Home">
@@ -55,25 +21,14 @@ export function Home({ data, setListPath, userId, userEmail }) {
 								setListPath={setListPath}
 							/>
 						))}
+					{
+						<CreateList
+							userId={userId}
+							userEmail={userEmail}
+							setListPath={setListPath}
+						/>
+					}
 				</ul>
-
-				<form onSubmit={handleSubmit}>
-					<h3>Create New Shopping List</h3>
-					<label htmlFor="newListName">Name Your List</label>
-					<br />
-					<input
-						type="text"
-						value={inputValue}
-						onChange={handleChange}
-						name="newListName"
-						id="newListName"
-						aria-label="Shopping List Name"
-						aria-required="true" // Indicates that this field is required
-					/>
-					<br />
-					<button aria-label="Create new shopping list">Create List</button>
-				</form>
-				<Toaster />
 			</div>
 		</>
 	);
